requirements.txt
python-docx==0.8.11
jieba==0.42.1
rank_bm25==0.2.2
aiohttp==3.9.1
tiktoken==0.5.2
tqdm==4.66.1
pandas==2.1.4
python-dotenv==1.0.0
openpyxl==3.1.2  # 用于Excel文件支持 